Updating with latest version from TodoMVC
As requested in #1044, here's the latest version of the Backbone.js
Todo app as rewritten by our project.

We started out with the 0.5 base and re-wrote it to cover some subtle
best practices we thought were important. Ours, like the current one
also uses the latest Backbone and jQuery 1.7.1. As part of the changes,
we also introduced two differences in the UX:

* When in edit mode, if a todo item is emptied and then blurred, the
item is removed. This contrasts with the current behaviour of the app
in the official repo at the moment which maintains the empty item in
place (albeit looking a little broken
http://addyosmani.com/gyazo/bbd4cd.png)

* We removed the tooltip occasionally seen when a user was trying to
add a new item. Having discussed this with developers frequently using
the Todo app as an initial point of reference, it was a consensus that
the notification didn't really offer that much value nor did it really
show anything that Bac